Rename legacy modules
