+++ src/legacy/modules/productsModule.js
